Added documentary and articles
Added a documentary on AI named Machine Learning: Living in the Age of AI by wired. Added a blog named "towards data science" for data science enthusiasts. I wanted to suggest that if we add two sections named " Research paper on AI" and "Blogs and Articles for AI" , it would be great. Please let me know about it. I have a ton of good research papers, blogs and articles to learn AI.