Update README.md
If you would like a change-by-change explanation, let me know and I will get it to you ASAP, but here is a summary of what I have done:

- Made all the reddit stuff follow the same format (instead of having `reddit.com/r/whatever` they all use `r/whatever`).
- Removed descriptions from reddit links, and moved them to the descriptions
- Removed many redundant descriptions
- Added a lot of correct number formatting (stopped sentences starting with numerals etc.). I will probably come back later and add some more of this.
- Tried to remove any partiality from descriptions (like "the best" or "a really good")

There are also a couple of entries commented out that either say the resource is not that great in the description (than why have it?) or did not fit in the category that they were put in. They can probably be revised and moved in the future.